backend/requirements.txt
# FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
authlib==1.3.2
itsdangerous==2.2.0

# Email
fastapi-mail==1.4.1

# Geospatial
geopy==2.4.1
shapely==2.0.6
numpy<2.0  # Shapely requires NumPy 1.x

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Security
slowapi==0.1.9
bleach==6.1.0

# Redis and WebSockets
redis==5.0.1
websockets==12.0

# Testing (optional)
pytest==7.4.4
httpx==0.26.0
